Added tech validation
If a goal couldn't be built or researched because its prerequisites
aren't met, its prerequisites will be pushed onto the goal list in front
of the goal.

For example, if at the start of the game we push Starport as a goal,
then Factory will be pushed in front of it. Since we're also missing
prerequisites for a Factory, Barracks will be pushed in front of
Factory.